Recursively resolve feature dependencies
This is needed to build the num crate, as a dependency of serde. The
problem is that one of the default features itself has feature
dependencies, like so:

  bigint = ["rustc-serialize", "rand"]
  default = ["complex", "rational", "bigint"]

Previously, this would have enabled "complex", "rational" and "bigint",
but not "rustc-serialize" or "rand", resulting in a build failure. With
this patch, resolution is performed recursively, and this build
succeeds.