Add a --no-download flag to avoid fetching crates
In theory, cargo-bootstrap would previously have worked in no-download
mode, as long as all of the versions installed are the latest versions
in the crates index. For use in a networkless environment, or a package
build system that needs to be usable without networking, it's sometimes
useful to have a set of pre-downloaded crates that satisfy dependencies
without being the latest versions available.

This commit adds a --no-download flag which forces cargo-bootstrap to
look only for locally downloaded versions of crates, and fail if a
dependency cannot be resolved that way.